requirements.txt
pytest-cov
pytest
fastapi[standard]>=0.113.0,<0.114.0
fastapi-pagination
jinja2>=3.1.4,<3.1.5
passlib[bcrypt]
pyjwt[crypto]